Add files via upload
Abstract
Downtime due to unscheduled maintenance is one of the prominent causes of inefficiency of operations in industries. Identifying and optimizing certain critical processes can often solve the problem to manageable levels. At the Inprotec plant, the air exhaust filter has been facing problems of clogging and moisture saturation due to an ill-designed flow diverter at the inlet, eventually causing a certain portion of filter bags to take an extra filtering load than the others. This project aimed to optimize the flow and try to eliminate the problem of clogging and identify a better location for fire hazard sensors—which are currently placed at positions that might not be optimum to ensure the desired functioning. Initially, the flow is analyzed without current design changes, and then with the current inlet diffuser manifold, by analyzing the two certain changes were done to the design to achieve close to homogenous flow through most filter bags.    
In the present study, the results of multiple modifications are analyzed based on flow behavior; qualitatively and quantitatively. Some modifications show promising perspective and will bring tangible value when implemented. CFD simulation showed that although information such as velocity distribution, flow behavior, and pressure distribution in the filter is tough to obtain by physical measurement, this data is relatively easily tractable using CFD investigation and is very beneficial for designing filters and optimizing retro-modifications.